Moved bind_to function to nat.rs
